+++ src/dns/implementation/types.ts
@@ -1,0 +1,4 @@
+export type Implementation = {
+  lookupTxtRecord: (domain: string) => Promise<string | null> 
+  lookupDnsLink: (domain: string) => Promise<string | null>
+}