+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Enable new EXPERIMENTAL public file system version 3.0.0 using rs-wnfs. Use this
 - Separates initialize into app and permissionedApp entrypoints
 - Make email optional at registration
 - Add reset option to bootstrapFileSystem
+- Add dependency-injected DNS lookup implementation
+- No longer uses a locally-shared IPFS client (was originally using a shared worker). This fixes various error messages you may have seen relating to CIDs.
+
 
 ### v0.32.0